+++ drivers/mstar2/include/mdrv_mstypes.h
@@ -78,7 +78,7 @@
     /// data type unsigned int, data length 4 byte
     typedef unsigned int                MS_U32;                             // 4 bytes
     /// data type unsigned int, data length 8 byte
-    typedef unsigned long               MS_U64;                             // 8 bytes
+    typedef unsigned long long          MS_U64;                             // 8 bytes
     /// data type signed char, data length 1 byte
     typedef signed char                 MS_S8;                              // 1 byte
     /// data type signed short, data length 2 byte
@@ -86,15 +86,15 @@
     /// data type signed int, data length 4 byte
     typedef signed int                  MS_S32;                             // 4 bytes
     /// data type signed int, data length 8 byte
-    typedef signed long                 MS_S64;                             // 8 bytes
+    typedef signed long long            MS_S64;                             // 8 bytes
     /// data type float, data length 4 byte
     typedef float                       MS_FLOAT;                           // 4 bytes
     /// data type pointer content
-    typedef size_t                      MS_VIRT;                            // 8 bytes
+    typedef unsigned long               MS_VIRT;                            // 8 bytes
     /// data type hardware physical address
-    typedef size_t                      MS_PHYADDR;                         // 8 bytes
+    typedef unsigned long               MS_PHYADDR;                         // 8 bytes
     /// data type size_t
-    typedef size_t                      MS_SIZE;                            // 8 bytes
+    typedef unsigned long               MS_SIZE;                            // 8 bytes
     /// data type 64bit physical address
     typedef MS_U64                      MS_PHY;                             // 8 bytes
 
@@ -126,13 +126,13 @@
     /// data type float, data length 4 byte
     typedef float                       MS_FLOAT;                           // 4 bytes
     /// data type pointer content
-    typedef size_t                      MS_VIRT;                            // 8 bytes
+    typedef unsigned int                MS_VIRT;                            // 8 bytes
     /// data type hardware physical address
-    typedef size_t                      MS_PHYADDR;                         // 8 bytes
+    typedef unsigned int                MS_PHYADDR;                         // 8 bytes
     /// data type 64bit physical address
     typedef MS_U64                      MS_PHY;                             // 8 bytes
     /// data type size_t
-    typedef size_t                      MS_SIZE;                            // 8 bytes
+    typedef unsigned int                MS_SIZE;                            // 8 bytes
 
 
     /// print type  MPRI_PHY
